+++ src/main/java/com/jwtauthentication/jwtauthsecurity/service/post/PostServiceInterface.java
@@ -9,9 +9,9 @@ import org.springframework.data.domain.Pageable;
 
 public interface PostServiceInterface {
     Page<PostResponse> getAllPosts(Pageable pageableable);
-    PostResponse getPostById(int id);
+    PostResponse getPostById(long id);
     PostResponse createPost(PostDto postDto);
-    PostResponse updatePost(PostDto postDto, int id);
-    String deletePost(int id, User user);
+    PostResponse updatePost(PostDto postDto, long id);
+    String deletePost(long id, User user);
     String getAuthenticatedUsername();
 }
